Migrate all tests into their own folder. Fixes #7
Also splits tests based on models.
