+++ tox.ini
@@ -51,8 +51,8 @@ deps =
 
 
 allowlist_externals = pytest
-python_files = tests.py test_*.py *_tests.py
-commands = pytest {toxinidir}/hubuum/tests.py
+# python_files = test/*.py
+commands = pytest {toxinidir}/hubuum/tests/
 
 #[testenv:docs]
 #extras = doc
